주거래 위치 api 생성 및 user entity float 변경

+++ src/api/user/entities/user.entity.ts
@@ -45,10 +45,10 @@ export class Users {
   @Column("int", { name: "gender", nullable: true })
   gender: number;
 
-  @Column("int", { name: "latitude", nullable: true })
+  @Column("float", { name: "latitude", nullable: true })
   latitude: number;
 
-  @Column("int", { name: "longitude", nullable: true })
+  @Column("float", { name: "longitude", nullable: true })
   longitude: number;
 
   @Column("varchar", { name: "location", length: 100, nullable: true })
